[191]: Unify GroupManager test files
